chore: Refactor ManageList component and move the forms for item names and purchase timelines, and logic to a separate AddItemForm component.
